Moved key namespace generation utils out of SecurityUtil
