+++ hbase-server/src/main/java/org/apache/hadoop/hbase/keymeta/KeyNamespaceUtil.java
@@ -1,0 +1,66 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.hadoop.hbase.keymeta;
+
+import org.apache.hadoop.hbase.client.ColumnFamilyDescriptor;
+import org.apache.hadoop.hbase.client.TableDescriptor;
+import org.apache.hadoop.hbase.regionserver.StoreContext;
+import org.apache.hadoop.hbase.regionserver.StoreFileInfo;
+import org.apache.hadoop.hbase.util.Pair;
+import org.apache.yetus.audience.InterfaceAudience;
+import org.apache.yetus.audience.InterfaceStability;
+
+/**
+ * Utility class for constructing key namespaces used in key management operations.
+ */
+@InterfaceAudience.Private
+@InterfaceStability.Evolving
+public class KeyNamespaceUtil {
+
+  /**
+   * Construct a key namespace from a table descriptor and column family descriptor.
+   * @param tableDescriptor The table descriptor
+   * @param family The column family descriptor
+   * @return The constructed key namespace
+   */
+  public static String constructKeyNamespace(TableDescriptor tableDescriptor,
+      ColumnFamilyDescriptor family) {
+    return tableDescriptor.getTableName().getNamespaceAsString() + "/"
+        + family.getNameAsString();
+  }
+
+  /**
+   * Construct a key namespace from a store context.
+   * @param storeContext The store context
+   * @return The constructed key namespace
+   */
+  public static String constructKeyNamespace(StoreContext storeContext) {
+    return storeContext.getTableName().getNamespaceAsString() + "/"
+        + storeContext.getFamily().getNameAsString();
+  }
+
+  /**
+   * Construct a key namespace from store file info.
+   * @param fileInfo The store file info
+   * @return The constructed key namespace
+   */
+  public static String constructKeyNamespace(StoreFileInfo fileInfo) {
+    Pair<String, String> tableNameAndFamilyName = fileInfo.getLink().getTableNameAndFamilyName();
+    return tableNameAndFamilyName.getFirst() + "/" + tableNameAndFamilyName.getSecond();
+  }
+}
